Preparing for dev release
Dynamic Image Resizer DONE! So we can now have the perfect performance on a image-based website. Also, only one or two commits away for the first pre-alpha release (code Harvard Panther) which is a massive milestone for the DSGNR project. If you read this commit message, well you are a true Scholar and a Gentleman